major crate structure overhaul
