+++ README.md
@@ -5,20 +5,6 @@ A work-in-progress runtime for [WASM-4](https://github.com/aduros/wasm4).
 
 ## Short Term Goals
 * [X] Implement all WASM-4 functions
-  - [X] `blit`
-  - [X] `blit_sub`
-  - [X] `line`
-  - [X] `hline`
-  - [X] `vline`
-  - [X] `oval`
-  - [X] `rect`
-  - [X] `text`
-  - [X] `tone` - sound slightly off
-  - [X] `diskr`
-  - [X] `diskw`
-  - [X] `traceUtf8`
-  - [X] `traceUtf16`
-  - [X] `tracef`
 * [X] Run W4 games (carts) on desktop platforms
 * [X] Embed wasmstation into standalone game executables
 * [ ] Offer support for different renderers (wgpu, sdl2)
@@ -27,7 +13,6 @@ A work-in-progress runtime for [WASM-4](https://github.com/aduros/wasm4).
 
 
 ## Mid Term Goals
-
 * Run W4 games on mobile and embedded platforms
 * Ability to compile games to run standalone without a WASM runtime (?)
 * allow input from non-standard sources via driver infrastructure (i2c devices on embeded)
