+++ core/src/framebuffer/text.rs
@@ -3,7 +3,6 @@ use crate::{Sink, Source};
 use super::blit_sub;
 
 const CHARSET_WIDTH: u32 = 128;
-const CHARSET_HEIGHT: u32 = 112;
 const CHARSET_FLAGS: u32 = 0; // BLIT_1BPP
 const CHARSET: [u8; 1792] = [
     0xff, 0xc7, 0x93, 0x93, 0xef, 0x9d, 0x8f, 0xcf, 0xf3, 0x9f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd,
