+++ desktop/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "wasmstation-desktop"
+version = "0.1.0"
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+
+
+[dependencies]
+wasmstation-core = { path = "../core/" }
+
+log.workspace = true
+bytemuck.workspace = true
+anyhow.workspace = true
+palette = "0.6"
+sdl2 = "0.35"
