+++ wasmer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "wasmstation-wasmer"
+version = "0.1.0"
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+
+
+[dependencies]
+wasmstation-core = { path = "../core/" }
+
+log.workspace = true
+bytemuck.workspace = true
+anyhow.workspace = true
+wasmer = "3.1"
