--- wasmstation/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "wasmstation"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# general
-log = "0.4"
-anyhow = "1.0"
-bytemuck = "1.12"
-num-traits = "0.2"
-byteorder = "1.4"
-
-# renderer
-sdl2 = "0.35"
-palette = "0.6"
-
-# backend
-wasmer = "3.0"
-
-# audio
-cpal = "0.14.2"
