--- wasmstation/src/backend/mod.rs
@@ -1,5 +1,0 @@
-//! Backends Which Run WASM-4 Carts
-
-mod wasmer;
-
-pub use self::wasmer::WasmerBackend;
